pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = "ScriptRAG — screenplay to Neo4j graph (self-healing extraction, Streamlit, metrics CLI)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "anthropic>=0.86.0",
    "instructor>=1.14.5",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "langgraph>=1.1.3",
]

[project.urls]
Repository = "https://github.com/kennygeiler/GraphRAG"
"Live demo" = "https://scriptrag.onrender.com/"
